# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "summary": "Criar novo pedido",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou estoque insuficiente"
          },
          "404": {
            "description": "Produto ou endereço não encontrado"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "summary": "Listar todos os pedidos (admin/gerente/funcionário)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "CONFIRMED",
                "PROCESSING",
                "READY_FOR_PICKUP",
                "OUT_FOR_DELIVERY",
                "DELIVERED",
                "CANCELLED",
                "REFUNDED"
              ],
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ONLINE",
                "IN_STORE"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/my-orders": {
      "get": {
        "operationId": "OrdersController_findMyOrders",
        "summary": "Listar meus pedidos",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "CONFIRMED",
                "PROCESSING",
                "READY_FOR_PICKUP",
                "OUT_FOR_DELIVERY",
                "DELIVERED",
                "CANCELLED",
                "REFUNDED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedidos do usuário retornados com sucesso"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "summary": "Obter detalhes de um pedido",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido retornado com sucesso"
          },
          "403": {
            "description": "Sem permissão para ver este pedido"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "summary": "Atualizar status do pedido",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso"
          },
          "400": {
            "description": "Transição de status não permitida"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "summary": "Cancelar pedido",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido cancelado com sucesso"
          },
          "400": {
            "description": "Pedido não pode ser cancelado"
          },
          "403": {
            "description": "Sem permissão para cancelar este pedido"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/{id}/tracking": {
      "get": {
        "operationId": "OrdersController_getTracking",
        "summary": "Rastrear pedido",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados de rastreamento retornados com sucesso"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Supermercado API",
    "description": "API completa para plataforma de supermercado - Venâncio Aires, RS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "uuid-do-produto",
            "description": "ID do produto"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Quantidade do produto",
            "minimum": 1
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre o item"
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Itens do pedido",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          },
          "fulfillmentType": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ],
            "example": "DELIVERY",
            "description": "Tipo de entrega"
          },
          "deliveryAddressId": {
            "type": "string",
            "example": "uuid-do-endereco",
            "description": "ID do endereço de entrega (obrigatório para DELIVERY)"
          },
          "couponCode": {
            "type": "string",
            "example": "CUPOM10",
            "description": "Código do cupom de desconto"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre o pedido"
          }
        },
        "required": [
          "items",
          "fulfillmentType"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PROCESSING",
              "READY_FOR_PICKUP",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED"
            ],
            "example": "CONFIRMED",
            "description": "Novo status do pedido"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre a mudança de status"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}